requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.2
Werkzeug==3.1.3
WTForms==3.2.1
Pillow==11.2.1
python-dotenv==1.1.0
python-slugify==8.0.4
psycopg2-binary==2.9.10
SQLAlchemy==2.0.28
alembic==1.13.1
requests==2.32.3
bleach==6.1.0
html2text==2024.2.26
email_validator==2.1.1
markdown==3.5.2
gunicorn==21.2.0
markupsafe==2.1.5
Jinja2==3.1.3
click==8.1.7
itsdangerous>=2.2.0
